politeping/requirements.txt
fastapi
uvicorn[standard]
httpx
PyYAML
python-dotenv